+++ far/02_href_regs.py
@@ -8,11 +8,9 @@ import json
 
 # This could have been anywhere on acq.gov
 srch = 'https://www.acquisition.gov/dears/part-970%E2%80%94doe-management-and-operating-contracts'
-
 # Open the url and save it as an html object
 resp = ul.request.urlopen(srch)
 html_res = resp.read()
-
 # Turn it into html and parse out the content
 soup = bsp(html_res, 'html.parser')
 rt = soup.find('div', class_ = 'reg-container clearfix')
@@ -22,26 +20,22 @@ rt = rt.find_all('a')
 # Remove all its contents before writing anything, but only if it exists
 jname = os.path.basename(__file__).split('.')[0]
 jname = 'json/' + jname + '.json'
-
-if path.exists(jname):
-  open(jname, 'w').close()
+if path.exists(jname): open(jname, 'w').close()
 
 with open(jname, 'w', encoding = 'utf8') as jf:
   jf.write('[')
-
   # Take-out the attributes
   for i in rt:
     reg_abb = i.get_text()
     reg_abb = reg_abb.strip()
     reg_name = i.attrs['href']
     reg_name = reg_name.strip()
-    
+    # Create dictionary to start adding values
     d = {
          'reg': reg_abb,
          'href': reg_name
         }
     json.dump(d, jf, indent = 2)
-    
   # Add closing bracket to signify the end
   jf.write(']')
 
@@ -62,10 +56,3 @@ with open(jname, 'w', encoding = 'utf8') as jf:
 print('Finished pushing data to ' + jname)
 
 
-
-      
-      
-      
-      
-      
-
